requirements.txt
# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Google Generative AI
google-generativeai>=0.3.0

# Data Visualization
plotly>=5.17.0

# Image Processing (for Aadhaar card processing)
opencv-python>=4.8.0

Pillow>=10.0.0

# HTTP Requests
requests>=2.31.0

# Other utilities
joblib>=1.3.0
